se realizo un login con roles de acceso

+++ frontend/src/app/app.ts
@@ -1,22 +1,29 @@
 import { Component, OnInit } from '@angular/core';
-import { NgFor } from '@angular/common';
+import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { UsuariosService, Usuario } from './usuarios';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [NgFor, FormsModule],
+  imports: [CommonModule, FormsModule],
   templateUrl: './app.html',
   styleUrl: './app.css',
 })
-export default class App implements OnInit {
+export class App implements OnInit {
   usuarios: Usuario[] = [];
 
   nombre: string = '';
   password: string = '';
   rol: string = '';
-  usuarioSeleccionado: any = null;
+
+  loginNombre: string = '';
+  loginPassword: string = '';
+
+
+  usuarioSeleccionado: Usuario | null = null;
+
+  usuarioLogueado: Usuario | null = null;
 
   constructor(private usuariosService: UsuariosService) {}
 
@@ -35,9 +42,50 @@ export default class App implements OnInit {
     });
   }
 
-  crearUsuario(): void {
+
+
+  login(): void {
+    const datosLogin = {
+      nombre: this.loginNombre, password: this.loginPassword
+  };
+    console.log('Enviando Login', datosLogin);
+
+    this.usuariosService.login(datosLogin).subscribe({
+      next: (respuesta ): void => {
+        console.log('Respuesta Login: ', respuesta);
+        if (respuesta) {
+          this.usuarioLogueado = respuesta;
+          this.loginNombre='';
+          this.loginPassword='';
+          alert('login correcto')
+        }else{
+          alert("Usuario o contraseña incorrectos");
+        }
+      },
+      error: (err: unknown):void => {
+        console.error('Error en login', err);
+        alert("Error al iniciaar sesion");
+      }
+    });
+  }
+  logout(): void {
+    this.usuarioLogueado = null;
+    this.usuarioSeleccionado = null;
+  }
+  esAdmin(): boolean{
+    return this.usuarioLogueado?.rol === 'adminastrador';
+  }
+  esEmpleado(): boolean{
+    return this.usuarioLogueado?.rol === 'empleado';
+  }
+  crearUsuario():void{
+    if(!this.esAdmin()){
+      alert('Solo el administrador purde crear usuarios');
+      return;
+    }
+
     if (!this.nombre || !this.password || !this.rol) {
-      alert('Todos los campos son obligatorios');
+      alert("Todos los campos son obligatorios");
       return;
     }
 
@@ -61,32 +109,62 @@ export default class App implements OnInit {
       },
     });
   }
+
   eliminarUsuario(id: number): void {
+    if (!this.esAdmin()) {
+      alert('solo el administrador puede eliminar usuarios');
+      return;
+    }
     this.usuariosService.eliminarUsuario(id).subscribe({
-      next: () => {
+      next: ():void => {
         console.log('Usuario eliminado');
         this.cargarUsuarios();
       },
-      error: (err: unknown) => {
+      error: (err: unknown):void => {
         console.error('Error al eliminar usuario', err);
       },
     });
   }
-  seleccionarUsuario(usuario: any){
-    this.usuarioSeleccionado = {...usuario}
-  }
-  cancelarEdicion(): void {
+  logOut(): void {
+    this.usuarioLogueado = null;
     this.usuarioSeleccionado = null;
-    alert("Edicion cancelada")
   }
-  actualizarUsuario(){
-    this.usuariosService.actualizarUsuario(this.usuarioSeleccionado.id, this.usuarioSeleccionado)
-      .subscribe(() =>{
-        this.cargarUsuarios();
-        this.usuarioSeleccionado = null;
-      });
 
+
+  seleccionarUsuario(usuario: Usuario):void {
+    if (!this.esAdmin()){
+      alert("solo el administrador puede editar usuarios");
+      return;
+    }
+    this.usuarioSeleccionado = { ...usuario };
   }
+
+  actualizarUsuario():void {
+    if(!this.esAdmin()){
+      alert('solo el administrador puede actualizar usuarios');
+      return;
+    }
+    if (!this.usuarioSeleccionado){
+      return;
+    }
+    if (!this.usuarioSeleccionado.nombre || !this.usuarioSeleccionado.password || !this.usuarioSeleccionado.rol){
+      alert("Todos los campos del usuario a editar son obligatorios");
+      return;
+    }
+    this.usuariosService.actualizarUsuario(this.usuarioSeleccionado.id, this.usuarioSeleccionado)
+      .subscribe({
+        next: () => {
+          this.usuarioSeleccionado = null;
+          this.cargarUsuarios()
+        },
+        error:(err: unknown) => {
+          console.error('Error al actualizar usuario', err);
+        }
+      });
+   }
+    cancelarEdicion(): void {
+    this.usuarioSeleccionado = null;
+   }
 }
